Move button to closed sprints badge
This is part of story #14179 Display sprints in release

Now the button to close sprints badges is moved to the left (between total sprint badge and closed sprint badge).
You can hover this 2 badges to display a closed arrow in the button (at the left).

How to test:
	- Have some releases
	- Have some sprints

Expected results:
	- If there are only closed sprints
		- You can't click on badge of total of closed sprints (and the badge is disabled color)

	- Else, you can click on the sprint badge to display the number of closed sprints
		- You have now a button to the left to close the badge of closed sprints

Change-Id: I8ecf98b3cfe9f870c70d7bf927003e0d6209e9d5