Unit tests can fail dependending of the execution order due to an incorrect casing of a classname
While PHP is not sensitive to the casing of classnames, Composer autoloaders
are. If we first create a class from a mock with an incorrect casing the
inheritance tree will be incorrect for this class and this can generate
errors for the rest of the test suite.

Part of request #14150: Get rid of SimpleTest

Change-Id: I9e67053873ac159b35668710f8cf03b9937599f0