Mass change for manage "en masse" backlog operations
This is part of story #14000 manage "en masse" backlog operations

A new action is proposed in masschange of artifacts that are in a top
backlog tracker. This action allows to add to top backlog, remove from top backlog, do nothing about
selected artifacts in the project's top backlog.

Add and remove are done in best effort, which means that nothing must be
done for artifacts that are already planned in a submilestone.

Change-Id: I458641b04726335150cdaf81d56acaefe89a44bc