request #14330: Under some conditions files can not be added to a FRS package
On CentOS 7 files pushed through the public FTP can not be added to the
release. perl-suidperl is not available on CentOS/RHEL 7 so the fileforge.pl
script is not executed by the correct user.

Ultimately the fileforge.pl script should be eradicated but for now, this
contribution only transforms the setuid bit that was set on this script into
a sudoers file.

To test it locally without rebuilding your dev Docker image, you will need
to deploy the new sudoers.d config file manually.

Change-Id: I46613b00635890e0e54ae5a92403be695edd9fcf