XML import of explicit backlog content
This is part of story #14000 manage "en masse" backlog operations

Top backlog content in explicit mangement can be added with REST project
XML import:

<top_backlog>
  <artifact artifact_id ="125"/>
</top_backlog>

It skips unkown artifacts and artifacts that are not in top backlog
trackers.
It does not import anything if the project does not have plannings.

Undone work is logged in /var/log/tuleap/project_xml_import_syslog

Change-Id: I9742fbb87560f70dd9f2781b6ea9f017fabc41b9