Introduce the notion of scopes on personnal access keys
Each personal access key now has scopes associated with it.
This will allow users to explicitely choose what the access key is
going to be used for.

The only visible change from this contribution is that the scopes
associated with an access key can be seen in the response given by
the REST endpoint GET /users/:id/access_keys.

Part of story #14023: access Git repositories with a personal access key

Change-Id: I81f23649f8f1bad43bb7f2f8fff4b2130e15513f