Display from an other project card
part of story #8283 Enhance project creation user experience

How to test:
- have a project admin of several projects
- You should be able to choose project in the drop down and use it as template

This is done in best effort:
- I do not have mockup for spinner, I used the standard spinner for now
- If any error happens during the load of project, it's displayed as an alert in the card

Rest route does not provide description, so unlike mockup it's not displayed in the drop down
Service modal will be added later.
Rest route does not provide project type, so if user is member of a template, it will be displayed twice:
once in company template, one in drop down => this will be corrected later.

Change-Id: I8e5fc7af1e822bd48383627fc607be2bfa041cd6