No drop below the add form
When add form is display with opacity:0, the card can be dropped
below. This produces unexpected behaviors, like having two drop zones:
one before the form, one below. This is too confusing for the end user.

In order to remove the issue, we don't display anymore the form. This
forces us to remove the opacity transition, but this is needed for a
better user experience.

Part of story #14249: Taskboard technical improvements

Change-Id: I2e1ef7d663fe72b4d429e4275c7afda9d1213436