Access keys can be created with a scope 'write:git_repository'
For now, this scope does nothing but ultimately with it users will be
able to do Git operations. However, with an access key having only a
scope allowing to do Git operations it is not possible to use the REST
APIs.

To test, you can use the REST endpoint POST /access_keys

Part of story #14023: access Git repositories with a personal access key

Change-Id: Id8ed873f30fc828f7230d92b7723614bd09638a5