Switch back from explicit backlog to the magical one
This is part of story #14000 manage "en masse" backlog operations

It's now possible to switch back from explicit backlog management to the
open and unplanned collection.

"Unplaned" option selected in a Tracker report is automatically switched
to "Any" when accessing the report for the first time after the switch
occurs. No feedback is displayed yet.

Change-Id: I2e415236caaaf2a28c25daa762fc65e981654f64