Access keys with the appropriate scope can be used to do Git over HTTP operations
To test you can create an access key via the REST endpoint POST /access_keys
with the scope "write:git_repository".

To be compatible with the most tools, basic authentication is used.
The user is suppose to provide a username and an access key instead ot its
password. When a valid access key is given with the incorrect username the
access key usage will be logged (it means the access key is known) but the
access will be denied.

Part of story #14023: access Git repositories with a personal access key

Change-Id: I12fa0ec276dcaf9e3c2d0c9a9f598bd0b2bbd921