request #14353: Build fail after a 'npm ci' at the source root
It seems there is difference (maybe a bug?) in the generated
node_modules tree after a `npm ci` and `npm install`. The tree
generated with `npm ci` can not be used to build Tuleap.

To workaround the issue, the usage of `npm ci` in the Makefile
target is removed . It was initially used to workaround another
bug/missing feature anyway and our test pipeline make sure the
lockfiles are in sync with our package.json files.

This contribution also updates the root npm lockfile that seems
to have been "semi-broken" by one the `npm dedupe` actions done
with the last month dependency updates (doing `npm ci` followed
by `npm install` changes the lockfile, that should not happen)

Change-Id: I6bd46295ab1d5fe327b147756f4f8a38620c636d