request #14372: Empty tracker template should be empty
As of today the XML template that is available to create an empty
tracker already contains some configuration: it contains fields & report
& renderers.

The proposal here is to first remove the fields and their permissions.

Removal of report from this template is more complicated because there
is no decent empty state when there is no report & renderer. At least
the graphontracker renderer can be removed.

Change-Id: I82dfce6bfa46d65b38dc9cb2be5bf42e2b2847db