Move the last Core stylesheets to Webpack to build the CSS
Everything is supposed to work as before.
This drops support for "npm run sass". There is only "npm run build"
now.

All gulp code related to SCSS work is also dropped.

I've chosen to build once the CSS for select2 and just include it in
sources. Given that's what we're doing for other vendor deps like
'atwho' or bootstrap, we can live with it. Besides, as long as it's not
an npm dependency, it's not going to change in those sources...

A couple of rules were removed from "style.scss": they referenced images
that were deleted in 027e5d7c61a23b5013d8fe103dcb553e9fb6058b in 2016.
They belonged to the STTab theme. I think it's safe to drop the rules.

Part of request #13299: Move all SCSS/CSS build process to Webpack

Change-Id: Id53c8d9b3002b30c2befbf8a5510290c756807f0