request #14397: API Explorer uses Swagger-UI v3
The API Explorer has been transformed into a plugin. Until now the API
Explorer was sort of a part of the Tuleap core that could be present or
not. This was manageable because the API Explorer was only static files,
this is not the case anymore.
To ease the transition the new plugin API Explorer is enabled by
default on all existing instances.

The page is always accessible as some REST endpoints can be queried
as anonymous even when the instance does not allow anonymous users.

To test, do not forget to redeploy nginx configuration since we have
forced to do some changes.

Change-Id: Id1c30b717a68edb0cb93b3838e5ff3bfefdb3489