request #14391: Compress JSON payloads
JSON payloads are not compressed, they should be as it can
makes things a bit faster for clients interacting with the
REST API.
To test you will need to redeploy the nginx configuration.
Also note, that if you tets behind a proxy the proxy might
decide to not compress it.

Change-Id: Id15b698e28105fcb2a7df82abcbebf40be426733