request #14390: Contributor semantic can be a radio or checkbox
Contributor semantic was done at a time where radiobutton or checkbox
types of list were inexistant. It seems that when those types have been
added to the codebase, we somewhat forgot to update the contributor
semantic edition page.

Have a tracker with only one list field: a selectbox.
Set the contributor semantic to this selectbox field.
Edit this selectbox field to convert it to a radio button.
Go back to the contributor semantic page. You should be able to edit the
contributor semantic.

Change-Id: I6abb24032f1ba05731445699757af686cb0d2b17