Inject assign to information in DOM
Have a tracker without any contributor semantic.
Load the taskboard.
The source of the page show assign_to_field: null.

Have a tracker with a contributor semantic on a non updatable field.
Load the taskboard.
The source of the page show assign_to_field: null.

Have a tracker with a contributor semantic on an updatable field.
Load the taskboard.
The source of the page show assign_to_field: { id: <int>, is_multiple:
<bool> }.

This is not used yet. This will be useful for later.

Part of story #14151: Update the assigned-to semantic of a card

Change-Id: Ib9238d82d32c349aecbe4fee6a92c336f1eae2f9