story #14392 Don't leak sub-milestones label
In ProjectMilestones widget, any users can see sprints badge.
But, if the user is not allowed to access the sprints tracker, he will see the tracker's label in the widget anyway.
It's a data leak.

Now, if the user can't access the tracker, he will not see the tracker's label.

How to test:
	- Have 2 users:
		- One can see all sprints
		- One can't see any sprints (or only his sprints)

Expected results:
	- With the one can see all sprints
		- No expected changes
		- The number of sprints displayed is the total of sprints

	- With the other user
		- If the user can't see any sprints
			- No sprint badge display
		- If the user can see some sprints
			- The number of closed sprints can be 'false'
			  This number is get by Rest API and can be change between the real number (get by API)
			  and the number of sprints can see the user. (This is a tradeoff: https://docs.tuleap.org/user-guide/rest/collections.html)

(In plugins/projectmilestones/include/Widget/ProjectMilestonesPresenterBuilder.php, it's the same behaviour
like in AD. If there are more 1 subplanning, only the first child is taken).

Change-Id: I2291235175feb06985bd1171fa933140c2e3c2de