Add datamodel and perform post action
This is part of story #14096 have a post action to add "Scrum Top Backlog" flag

The datamodel (database structure) for a new "add to top backlog" post
action has been added.

You have to add manually the post action in the MySQL table
named plugin_agiledashboard_tracker_workflow_action_add_top_backlog

Once added, when an artifact do the transition, then it's added in the
explicit top backlog if the project is configured to use the explicit
backlog.

Nothing must occurs if the artifact is either already in the top backlog
or already planned in a submilestone.

No feedback is displayed yet, it will be handled in a dedicated commit.

Change-Id: I10a8c0126889be63b68ecc1de3c9ed3520c89dd0