request #14398 Project creation service check should not rely on project 100
Enable the new UI with: $can_use_new_project_creation=1; in your local.inc

How to reproduce:
Disable Agiledashboard service from project 100
Tuleap templates: scrum, kanban and Tuleap alm are not available

Now if you want to test:

Happy path scénario:
Go on project/new url
Expected result: Tuleap templates: scrum, kanban and Tuleap alm should be available

All plugins are not available in platform scenario:
Disable Agiledashboard in plugin section
Expected result : Previous template must not be available

Restrict plugin scenario:
Restrict the agiledashboard plugin
Expected result: Previous template are available
Please note that restricted plugins are not taken into account in this changeset for those main reasons:
Not all plugins implement correctly the notion of restricted plugins (but it is another issue that should be handled separately).
As of today, we don't know how to handle a restricted plugin, mainly because we don't really know how this feature is used:
It is used for progressive deployment of new features. In that case we should probably enable the underlying templates and after creation, add the new project to the list of allowed project to use the plugin.
It is used to restrict the access to a feature. In that case we should not allow the selection of the underlying templates.
Until we have a better view on which direction to take for those restricted plugins, the best-effort approach has been taken: templates are available, but new project won't have access to the plugin.

/!\ this patch need hook of TTM to be tested with I9f7226f08e0acde70ae45339b45ebfdc3313e6ca
 => without TTM patch Tuleap ALM project won't be available

/!\ Only services used by tuleap templates listen the hook, and only core used by templates are declared:
we do not want the hook to have proftpd or forum, we do not want to instantiate this legacy services in templates provided by Tuleap.




Change-Id: Ib29d3c060d00c72b655f95cb3b926a4149aea2e3