Display user edit icon in edit mode
Switch a card to edit mode. An icon is displayed to inform that the list
of assignees can be edited. The icon is different if the existing
list of assignees is empty or not.

Icons appear only if the corresponding assignee field can be edited.

Part of story #14151: Update the assigned-to semantic of a card

Change-Id: I9563b0566dd18d6774edfb96086558b3787cee08