I can choose assignee for a card
Have a story tracker with a multiselectbox as assign to, and set
contributor semantic accordingly. Field must be updatable.

Have a children task tracker with a selectbox as assign to, and set
contributor semantic accordingly. Field must be updatable.

Switch a story to edit mode. Click on assignees. You can choose more
than one people to add/remove from the list. Cancel the edition.

Switch a task to edit mode. Click on assignee. You can choose only one
assignee. Cancel the edition.

If server takes time to retrieve the users, then a spinner is displayed
so that user knows that action is ongoing.

Note: user (de)selection is not saved for now.
Note: users are always retrieved. We should save resources by making
only one call per field. This can be done in a dedicated changeset.

Part of story #14151: Update the assigned-to semantic of a card

Change-Id: Id7acb92c2dba69930e38b26edb9b18b0d6d56363