request #14406: Crash when accessing a project having its first service displayed in a iframe
Reproduction scenario:
 1. In a project have a service displayed in a iframe (you might need
to play directly with the DB to achive that)
 2. Disable the summary project
 3. Set the rank of the service displayed in the iframe to 1

The issue is present for a while and the use case is narrow enough
to just disable and remove completly this feature. The feature cannot
be used at all on recent instances since it is no more possible to create
a service displayed in a iframe.

Change-Id: Ia0c92495030441c71ec7c9b3daee469d29c2b5b6