Group update of the SVN access files of a project when possible
When there is multiple changes affecting ugroups in the same project
Tuleap should try to group the refresh of the SVN access files. This
operation can be a bit long in large projects so in situation where
there is lot of ugroup changes (e.g. automation, nightly synchro…)
it can slow (a lot!) the processing of default system event queue.

To avoid that, a new system event that deal only with the update of the
SVN access file is created. This event is queued only when there is not
already an event pending for a project.

Close request #14411: Refresh SVN access files only once when multiple
system events need it

Change-Id: I0b4d99c1674e04c104bef4102aec1f049b200fe5