Preselect users
Given a card has already some assignees,
When I switch the assignees to edit mode,
Then the users should be preselected.

Note: The selectbox is using a select2 multiple with a max selection of
1. This is a temporary solution to avoid having the first value
autoselected when we open the list, which is strange. A dedicated commit
will deal with this issue.

Bonus: I can not switch anymore the assignees to edit mode if the card
itself is not in edit mode.

Bonus2: No need to have a minimum input length, we can suggest instantly
the list of the team so that user can select people without having to
leave the mouse.

Bonus3: The placeholder is displayed for both selectbox and
multiselectbox.

Part of: story #14151 Update the assigned-to semantic of a card

Change-Id: I6ee7b3f44b3307ee62fb98c0d570f45f38f22e9e