On switch from explicit to magical, remove post actions
This is part of story #14096 have a post action to add "Scrum Top Backlog" flag

When switching back from explicit to magical, all the tracker workflow
post-action "AddToTopBacklog" will be removed.

This behavior is explained in the warning message under the switch
button.

The name of the post-action, AddToTopBacklog, has not been yet
translated in french, so I kept it as is in the feedbacks.

Change-Id: If22ac2afe4885c6acee1acfbb909817a1169d40c