When the card is saved, new assignees are also saved
Update the assignees of a card, and save the card.
Refresh the page. The new assignees are kept.

Note: during the save of the card, there is a small moment when the card
still displays the previous value of assigned to. »Then the card is
refreshed so it might be transparent for users with low latency.) This
issue will be handled in a dedicated commit.

Part of: story #14151 Update the assigned-to semantic of a card

Change-Id: I761aa3dbf3fd6af421e1c70d1e6492f0bd8b8381