story #14131 : Refactor user suspension code
All user suspension and user suspension notification functionality is moved to UserSuspensionManager
User suspension methods from UserDao are moved to UserSuspensionDao (with adaptation to the latest DataAccessObject). There is already logic in some of the dao methods in the legacy code. Since the goal for now is to just refactor, the same structure is kept.

Change-Id: I748a8e632afa35ba401b8dc0ae735ef58d2d5872