Remove superfluous UserAjaxRepresentation
UserAjaxInterpretation was needed because legacy ajax call to get
possible users was not returning `display_name` property (it uses the
generic `label` instead, that is common to all ListValue).

We were forced to have three interfaces for user:
* User: the assignee of a card
* UserAjaxRepresentation: the possible assignees
* UserForPeoplePicker: mixin of UserAjaxRepresentation and DataFormat so
that select2 can display possible assignees in the dropdown & selection.

With this commit, now legacy ajax call returns an additional property
`display_name` so that the result are compliant with taskboard's User
interface. This reduces the number of interfaces for user:
* User: still the assignee of a card
* UserForPeoplePicker: more or less a mixin of User and DataFormat for
select2.

No functional changes (except in the ajax call response).

Part of: story #14151 Update the assigned-to semantic of a card

Change-Id: I80dcdeeb5a168e5293e69746f36f8d12d8ea8dd7