Adapt packaging to build version flagged as Enterprise
It is now possible to include the ENTERPRISE_BUILD file
located at the sources root. This file is used to determine
if the version currently running is an Enterprise build or not.

Note that if you use the tools/rpm/build_and_run_packages.sh script
to test this, the file must be present in the "clean sources" directory
that get created before the preparation/build phase.

This is part of story #14112: know which Tuleap is running

Change-Id: I5aff5cb7a63d1dd7c56ee77227446e36075bafc4