Update card before its refresh
If refresh of the card takes times to proceed, then the new assignee are
not displayed (whereas the new label is). In order to remove this
confusion, the new assignee should be displayed as soon as possible.

You can test that with a `sleep()` in
`\Tuleap\Taskboard\REST\v1\TaskboardCardResource::getId()`

Implementation notes
--------------------

We could have achieved the same result without having to change the
whole `new_assignees_ids` chain, because the possible assignees are
cached by the store and the new assignees can therefore be retrieved by
card-actions. However this would force us to declare another payload for
card-mutation. As of today `card-mutation:finishSavingCard()` and
`card-actions:saveCard()` are sharing the same UpdateCardPayload and
IMHO we should stick to it for now. Furthermore using real objects
instead of primitive is not that bad.

PeoplePicker still return list of ids, because it does only know about
UserForPeoplePicker, not User. We could return a list of
UserForPeoplePicker and let CardAssignees do the conversion to list of
User, however I don't think it is really worth it.

Part of: story #14151 Update the assigned-to semantic of a card

Change-Id: Ib4e4c9af7ca1311de2ab5a4a7d65cb6098ffada9