request #14431: Limit the number of times an async event can be re-queued
When the processing of an async event crashes/timeout the event is added
back in queue which is great for reliability in case of a temporary failure.
However, if the crash/timeout is due to something more permanent retrying indefinitely
does not help and it can be a source of issues (e.g. re-sending the same notification over and over).

Change-Id: Ia5232b13c571376827f7d08f26a65c571739d020