Manage the "build" of the fat|subsets* core JS files with Webpack
A plugin is introduced to manage the concatenation and global inclusion
of the multiples JS files into one. This is needed because in the normal
build process Webpack will wrap the code into a new scope, unfortunately
those legacy files expect to be in the global scope.
Also a "normal" Webpack config cannot have a empty 'entry' property so
to bypass that and have a Webpack config that only deals with the legacy
files a plugin has been created.

Part of request #14384 Move all javascript build process to Webpack

Everything is supposed to work as before.

Change-Id: I7fe79d58a8ffa2fc8d987e259e9c9f3ac3f5f4e8