XML import of AddToTopBacklog post action
This is part of story #14096 have a post action to add "Scrum Top Backlog" flag

You can now add an AddToTopBacklog post action with the tracker XML
import:

<postactions>
  <postaction_add_to_top_backlog/>
</postactions>

Only one tag is allowed. It must work in both simple or advanced
workflow mode.

XML Import must not be done if a post action is defined and cannot be
imported in both tracker or full project import.

Better error messages displayed when the pre checks are not OK in
project XML import will be handled in a dedicated commit.

Change-Id: I51f9ef37d24f8b31b7f5e7500b967f47102b99a8