story #14165 Use route rest artifact/$id to get burndown data instead of milestone/$id/burndown
In Project Milestones widget, we use a new route rest to get burndown.
Now, we use artifact/:id, instead of milestone/:id/burndown.

This change allows to get burndown and burnup with a single request.
In the futur, Project Milestone will display the burnup, and there will be not necessary to use a new request.

No expected changes.

Change-Id: Ic691c4f9beb18a556c1b15b53b24b925b5daa8fc