Introduce the new post action option "Add to backlog" in the post action modal
Part of story #14096 have a post action to add "Scrum Top Backlog" flag

The patch only introduce the display of the disabled option, there is no functional change.

The patch also introduce a way to import dynamically a vue component from an other plugin.
So how it briefly works ?
 1) The component is built with webpack as a JSONP and stored in the window global variable.
 2) A PHP event send the asset and the JSONP callback name to the tracker plugin.
 3) Script tags are built with the asset url to be able to retrieve them with the JSONP callback.
 4) We use the dynamic import in the target component.

Why use this mecanism?
With a webplack plugin[0] the dynamic import can be heandled more easly. BUT it does NOT work on IE 11, because this awesome browser does not support the ESM specs.

[0] https://github.com/purtuga/esm-webpack-plugin

Change-Id: Idebc38d956ce5dcbbd85f22a591a121d6cc236b7