Projects dropdown is below project categories select2
part of: story #14151 Update the assigned-to semantic of a card

I took the oportunity to re-align z-indexes of fly-over elements in
order to establish an actual hierarchy between those elements.

With this patch (in Burning Parrot) we have:

breadcrumb dropdowns < select2 field < navbar dropdowns < modal backdrops < modals

However select2 dropdown should be > modal so it is not hidden by the
modal. (check in kanban edit modal and pullrequest reviewers modal)

Change-Id: Ia823d4b0257cbc55cdab3d07f4b6d8044ee2bc62