Transform GitXmlImporterTest in a PHPUnit integration test
This test is transformed into a integration test instead of a unit test
because it needs access to some files deployed by the gitolite RPM.

The DB/Integration tests are now executed in an environnement that is
closer to production since the tests are now run as the codendiadm user
instead of a generic user.

One change to the production code has been introduced so it can
work without a privileged user.

This is part of request #14150: Get rid of SimpleTest

Change-Id: I397b0708ca3136ec86f68c9ec19ae310a71ec8a2