Remove animation of remaining effort
Part of story #14310 stop being bothered by the hidden add-in-place
editor when I drag and drop some cards

Since 557f8131ab44d3369bd46caf2d1203a8e35256f4 and the introduction of
our own library, it is now the browser that is reponsible for rendering
the "mirror" image that follows the pointer while dragging.

There is no longer a "gu-mirror" equivalent CSS class, and this
animation is no longer visible. We should drop it.

Change-Id: I7a3f01febf12f243fc718eaf3a7bfe41c911e28e