Make possible to introspect a personnal access key with a REST scope
A new REST endpoint is introduced to retrieve information about
an access key: GET /access_keys/{id}. "self" or a specific key ID
can be used for the {id} parameter. When "self" is given, information
about the key used to authenticate are returned.

This will allow external apps to determine easily if a key has the
required scopes.

Part of story #14019: automate jenkins job creation & run

Change-Id: I19f6c9a5fa13bd346817d5577650312b11d67bf8