Revert "Introduce the new post action option "Add to backlog" in the post action modal"
This reverts commit dbd1a274d5ba02799d4eb373f39039d7ebe8edef.

Part of story #14096 have a post action to add "Scrum Top Backlog" flag

There are several problems with this approach:

- AgileDashboard Vue components cannot import components from Tracker.
In this situation, it means they cannot use PostAction, while all other
PostActions components do use it, which creates inconsistency.
If an external component DOES import PostAction, it creates two versions
of it, and unexpected things start to occur. For example,
getComponentByName() will return things for the Tracker version and will
not for the AgileDashboard version. It is very disturbing for
developpers...

- This means we cannot use Vue slots in combination with external
components, and they cannot instanciate any component from their
importer. In other words, it's almost like two separate Vue apps.

- Typescript complains at every turn. This is because we now depend on a
Vue instance that is not present anywhere in AgileDashboard, but present
in Tracker. Typescript cannot know that we have $gettext or $store
available.

- Unit-testing is a pain too. In order to test, we need to install
another version of vuex, vue-test-tools, etc. Combined with Typescript,
it makes things difficult. It also means we can have different "test"
versions of Vue, vuex, etc. between importer and imported, but they will
share the same "prod" versions. Nothing can stop it.

- We have no choice but to check for undefined when declaring our
external components. Before "setExternalComponents" is called, the array
of external components is empty.

- The whole JSONP mechanism is very brittle. We are forced to create
script tags manually to control the order of execution. The only thing
tying the importer and the exporter together is the JSONP callback name,
so you must be sure to use the same name in the "library" config in
webpack and the function call in "window" in the importer.
Worse yet, we need to be 100% certain that the JSONP function exists on
window BEFORE the JSONP script is executed. Otherwise, error.
That means we are forced to build manual Promises that only resolve once
the JSONP callback has been called.
We would have used dynamic import() which is much simplet, except that we
need to support IE11 which does not have it.

For all these reasons, we should revert this mechanism and use a simpler
method.

Change-Id: I04765ad1fa2f4adeec4ce0f4c3129dfe156ad8ae